Default data reader function and tagged literals
Sometime it is not possible to know in advance which tagged readers
need to be registered when reading EDN data.

For example, I would like to read Splunk log lines that have been
produced by many different systems, and I do not know in advance which
data readers those system have defined.

However, it's still useful to read this data and maybe only use some
of the data, or pass the data on to other systems. Clojure 1.7 added
support for this via so called default data readers.

This PR implements a way to provide a default data reader function. It
is modeled after Clojure's implementation which is described here [1].

[1] https://insideclojure.org/2018/06/21/tagged-literal/