Remove all `:ast` values
parseedn-read never sees the full value of the `:ast` keys as defined
in `parseedn-test-data.el`. In fact, it doesn't know about it at
all. The only functions that see anything related to the `:ast` values
are `parseedn-reduce-leaf` and `parseedn-reduce-branch`. But they are
called back from `parseclj` when needed, to "reduce" (transform the
current AST node definition into the associated Emacs Lisp data type).

As commented in https://github.com/clojure-emacs/parseedn/pull/17#discussion_r1321814637,
they very likely are remnants of when parseedn and parseclj were the
same library.