Extract the logic of building the map to separate defuns
Before adding the code to support prefixed maps, the branch reduction
function looked more concise and higher-level. After adding the support
it seemed like it mixed levels of abstraction.

Extracting the nitty-gritty details of how to build the maps
(especially in the prefixed map case) should help in keeping the
levels of abstraction separated.